requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
opencv-contrib-python==4.1.1.26
Pillow==6.2.1
pyparsing==2.4.4
python-dateutil==2.8.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
urllib3==1.25.6
